Checkpoint: More moves of cf-specific code from store package
