+++ src/frontend/packages/cloud-foundry/src/cloud-foundry.module.ts
@@ -11,11 +11,6 @@ registerCFEntities();
     CloudFoundryStoreModule,
     // TODO: NJ split out anything lazy loaded into seperate module
     CloudFoundryComponentsModule,
-    // FIXME: STRAT-155 - Move cf effects into cf module
-    // EffectsModule.forRoot([
-    //   PermissionsEffects,
-    //   PermissionEffects
-    // ])
   ],
 })
 export class CloudFoundryPackageModule { }
