+++ src/frontend/packages/core/src/features/applications/deploy-application/deploy-application.component.ts
@@ -13,7 +13,7 @@ import { CFAppState } from '../../../../../cloud-foundry/src/cf-app-state';
 import { applicationEntityType } from '../../../../../cloud-foundry/src/cf-entity-factory';
 import { getCFEntityKey } from '../../../../../cloud-foundry/src/cf-entity-helpers';
 import { RouterNav } from '../../../../../store/src/actions/router.actions';
-import { selectApplicationSource, selectCfDetails } from '../../../../../store/src/selectors/deploy-application.selector';
+import { selectApplicationSource, selectCfDetails } from '../../../../../cloud-foundry/src/selectors/deploy-application.selector';
 import { selectPaginationState } from '../../../../../store/src/selectors/pagination.selectors';
 import { DeployApplicationSource, SourceType } from '../../../../../cloud-foundry/src/store/types/deploy-application.types';
 import { CfAppsDataSource } from '../../../shared/components/list/list-types/app/cf-apps-data-source';
