+++ src/frontend/packages/core/src/shared/components/add-service-instance/specify-user-provided-details/specify-user-provided-details.component.ts
@@ -19,7 +19,7 @@ import {
   userProvidedServiceInstanceEntityType,
 } from '../../../../../../cloud-foundry/src/cf-entity-factory';
 import { createEntityRelationKey } from '../../../../../../store/src/helpers/entity-relations/entity-relations.types';
-import { selectCreateServiceInstance } from '../../../../../../store/src/selectors/create-service-instance.selectors';
+import { selectCreateServiceInstance } from '../../../../../../cloud-foundry/src/selectors/create-service-instance.selectors';
 import { APIResource } from '../../../../../../store/src/types/api.types';
 import { IUserProvidedServiceInstance } from '../../../../core/cf-api-svc.types';
 import { safeUnsubscribe, urlValidationExpression } from '../../../../core/utils.service';
