+++ src/frontend/packages/store/src/types/routing.type.ts
@@ -1,4 +1,4 @@
-import { AppRoutingSate } from '../app-state';
+import { AppRoutingOnlyAppState } from '../app-state';
 
 export interface RoutingEvent {
   id: number;
@@ -24,9 +24,9 @@ export const defaultRoutingState: RoutingHistory = {
   currentState: null
 };
 
-export function getPreviousRoutingState(state: AppRoutingSate) {
+export function getPreviousRoutingState(state: AppRoutingOnlyAppState) {
   return state.routing.previousState;
 }
-export function getCurrentRoutingState(state: AppRoutingSate) {
+export function getCurrentRoutingState(state: AppRoutingOnlyAppState) {
   return state.routing.currentState;
 }
